requirements.txt
django-silver>=0.9
django-payu-ro==1.2.0
